url_inspector/requirements.txt
beautifulsoup4==4.11.1
Django==4.0.4
djangorestframework==3.13.1
requests==2.27.1
psycopg2==2.9.3
